requirements.txt
# Telegram Bot API
python-telegram-bot>=21.0

# Video downloader
yt-dlp>=2024.0.0

# Additional dependencies
certifi